FinalProject/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
SQLAlchemy
pymysql

langchain
langgraph
langchain-openai
langchain-chroma

requests
mysql-connector-python
passlib[bcrypt]
python-jose[cryptography]
html2docx
python-docx
beautifulsoup4
